+++ FileBaseContext.Tests/Serializers/JsonRowDataConverterTests.cs
@@ -411,7 +411,7 @@ namespace FileBaseContext.Tests.Serializers
                 [
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[]",
                 json);
         }
@@ -426,7 +426,7 @@ namespace FileBaseContext.Tests.Serializers
                     [],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                  "[\r\n  {}\r\n]",
                  json);
         }
@@ -444,7 +444,7 @@ namespace FileBaseContext.Tests.Serializers
                     ["String1"]
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": null\r\n  },\r\n  {\r\n    \"Value\": \"\"\r\n  },\r\n  {\r\n    \"Value\": \"String1\"\r\n  }\r\n]",
                 json);
         }
@@ -462,7 +462,7 @@ namespace FileBaseContext.Tests.Serializers
                     [10, 11],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": 0,\r\n    \"Nullable\": null\r\n  },\r\n  {\r\n    \"Value\": 10,\r\n    \"Nullable\": 11\r\n  }\r\n]",
                 json);
         }
@@ -480,7 +480,7 @@ namespace FileBaseContext.Tests.Serializers
                     [false, true],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": true,\r\n    \"Nullable\": null\r\n  },\r\n  {\r\n    \"Value\": false,\r\n    \"Nullable\": true\r\n  }\r\n]",
                 json);
         }
@@ -498,7 +498,7 @@ namespace FileBaseContext.Tests.Serializers
                     [10L, 11L],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": 0,\r\n    \"Nullable\": null\r\n  },\r\n  {\r\n    \"Value\": 10,\r\n    \"Nullable\": 11\r\n  }\r\n]",
                 json);
         }
@@ -517,7 +517,7 @@ namespace FileBaseContext.Tests.Serializers
                     [value, value],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": \"0001-01-01T00:00:00\",\r\n    \"Nullable\": null\r\n  },\r\n  {\r\n    \"Value\": \"2000-01-02T03:04:05.678\",\r\n    \"Nullable\": \"2000-01-02T03:04:05.678\"\r\n  }\r\n]",
                 json);
         }
@@ -539,7 +539,7 @@ namespace FileBaseContext.Tests.Serializers
                     [valueLocal, valueLocal],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": \"0001-01-01T00:00:00+00:00\",\r\n    \"Nullable\": null\r\n  },\r\n  {\r\n    \"Value\": \"2000-01-02T03:04:05.678+00:00\",\r\n    \"Nullable\": \"2000-01-02T03:04:05.678+00:00\"\r\n  },\r\n  {\r\n    \"Value\": \"2000-01-02T03:04:05.678-08:00\",\r\n    \"Nullable\": \"2000-01-02T03:04:05.678-08:00\"\r\n  }\r\n]",
                 json);
         }
@@ -559,7 +559,7 @@ namespace FileBaseContext.Tests.Serializers
                     [value, value],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": \"00:00:00\",\r\n    \"Nullable\": null\r\n  },\r\n  {\r\n    \"Value\": \"1.02:03:04.5670000\",\r\n    \"Nullable\": \"1.02:03:04.5670000\"\r\n  }\r\n]",
                 json);
         }
@@ -577,22 +577,22 @@ namespace FileBaseContext.Tests.Serializers
                     [(byte[])[0, 1, 2, 3]],
                 ]);
 
-            AssertStrings(
+            TestHelpers.AssertString(
                 "[\r\n  {\r\n    \"Value\": null\r\n  },\r\n  {\r\n    \"Value\": \"\"\r\n  },\r\n  {\r\n    \"Value\": \"AAECAw==\"\r\n  }\r\n]",
                 json);
         }
 
-        private void AssertStrings(string str, string json)
-        {
-            if (str.Length != json.Length)
-                Assert.Fail($"Strings have different length: {str.Length} != {json.Length}");
+        //private void AssertStrings(string str, string json)
+        //{
+        //    if (str.Length != json.Length)
+        //        Assert.Fail($"Strings have different length: {str.Length} != {json.Length}");
 
-            for (int index = 0; index < str.Length; index++)
-            {
-                if (str[index] != json[index])
-                    Assert.Fail($"'{str[index]}' != '{json[index]}' at {index}.");
-            }
-        }
+        //    for (int index = 0; index < str.Length; index++)
+        //    {
+        //        if (str[index] != json[index])
+        //            Assert.Fail($"'{str[index]}' != '{json[index]}' at {index}.");
+        //    }
+        //}
 
         private static string Serialize(
             IEnumerable<JsonColumnInfo> columns,
